CC-13-Controlar-los-botones

+++ src/Ventanas/Aulas.java
@@ -34,9 +34,16 @@ public class Aulas extends javax.swing.JPanel {
          //jtxtTipo.setEditable(false);
         CRUDEAulas objAulas = new CRUDEAulas();
         objAulas.mostrarAulas(jtblListaAulas,busqueda.getText());
-        objAulas.llenarComboBoxEdificios(ComboEdificio);
-        objAulas.llenarComboBoxTipo(ComboTipo);
-    }
+        objAulas.llenarComboBoxEdificios(comboEdificio);
+        objAulas.llenarComboBoxTipo(comboTipo);
+        
+        limpiarCampos();
+        botonesEditar();
+        textosEditar();
+        textosInicio();
+        botonesInicio();
+        botonesNuevo();
+        textosActivados();    }
     
     
         
@@ -51,8 +58,72 @@ public class Aulas extends javax.swing.JPanel {
     jtxtID.setText("");
     jtxtNombre1.setText("");
     jtxtCapacidad.setText("");
-   
+    //jtxtEdificio.setText("");
+    //jtxtTipo.setText("");
 }
+     public void botonesInicio() {
+        btnNuevo.setEnabled(true);
+        jbtnGuardar.setEnabled(false);
+        jbtnModificar.setEnabled(false);
+        jbtnEliminar.setEnabled(false);
+        jButton1.setEnabled(false);
+    }
+
+    public void textosInicio() {
+        jtxtID.setEnabled(false);
+        jtxtNombre1.setEnabled(false);
+        //jtxtEdificio.setEnabled(false);
+        jtxtCapacidad.setEnabled(false);
+        comboTipo.setEnabled(false);
+        comboEdificio.setEnabled(false);
+        
+
+    }
+
+    public void textosActivados() {
+        jtxtID.setEnabled(false);
+        jtxtNombre1.setEnabled(true);
+        //jtxtEdificio.setEnabled(true);
+        jtxtCapacidad.setEnabled(true);
+        comboTipo.setEnabled(true);
+        comboEdificio.setEnabled(true);
+        //jtxtTipo.setEnabled(true);
+        //jcmbCursos.setEnabled(true);
+    }
+
+    public void textosEditar() {
+         jtxtID.setEnabled(false);
+        jtxtNombre1.setEnabled(true);
+        //jtxtEdificio.setEnabled(true);
+        jtxtCapacidad.setEnabled(true);
+        comboTipo.setEnabled(true);
+        comboEdificio.setEnabled(true);
+    }
+
+    public void botonesNuevo() {
+        btnNuevo.setEnabled(false);
+         jbtnGuardar.setEnabled(true);
+        jbtnModificar.setEnabled(false);
+        jbtnEliminar.setEnabled(false);
+        //System.out.println("jokassa");
+        jButton1.setEnabled(true);
+
+    }
+
+    public void botonesEditar() {
+        int var = jtblListaAulas.getSelectedRow();
+        if (var != -1) {
+            btnNuevo.setEnabled(false);
+        jbtnGuardar.setEnabled(false);
+        jbtnModificar.setEnabled(true);
+        jbtnEliminar.setEnabled(true);
+        //System.out.println("jokassa");
+        jButton1.setEnabled(true);
+
+        }
+    }
+
+    
     
 
     /**
@@ -81,8 +152,10 @@ public class Aulas extends javax.swing.JPanel {
         jLabel8 = new javax.swing.JLabel();
         jLabel9 = new javax.swing.JLabel();
         jButton1 = new javax.swing.JButton();
-        ComboTipo = new javax.swing.JComboBox<>();
-        ComboEdificio = new javax.swing.JComboBox<>();
+        btnNuevo = new javax.swing.JButton();
+        jLabel12 = new javax.swing.JLabel();
+        comboEdificio = new javax.swing.JComboBox<>();
+        comboTipo = new javax.swing.JComboBox<>();
         jPanel2 = new javax.swing.JPanel();
         jScrollPane1 = new javax.swing.JScrollPane();
         jtblListaAulas = new javax.swing.JTable();
@@ -153,7 +226,7 @@ public class Aulas extends javax.swing.JPanel {
         jLabel7.setText("Edificio Ubicado");
 
         jLabel8.setFont(new java.awt.Font("Consolas", 0, 14)); // NOI18N
-        jLabel8.setText("Tipo");
+        jLabel8.setText("Tipo Espacio");
 
         jLabel9.setIcon(new javax.swing.ImageIcon(getClass().getResource("/images/escoba.png"))); // NOI18N
 
@@ -165,32 +238,48 @@ public class Aulas extends javax.swing.JPanel {
             }
         });
 
-        ComboTipo.setToolTipText("");
+        btnNuevo.setText("Nuevo");
+        btnNuevo.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnNuevoActionPerformed(evt);
+            }
+        });
+
+        jLabel12.setIcon(new javax.swing.ImageIcon(getClass().getResource("/images/usuario.png"))); // NOI18N
 
         javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
         jPanel1.setLayout(jPanel1Layout);
         jPanel1Layout.setHorizontalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
+                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
                     .addGroup(jPanel1Layout.createSequentialGroup()
+                        .addContainerGap()
+                        .addComponent(comboEdificio, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel1Layout.createSequentialGroup()
                         .addContainerGap()
                         .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                             .addComponent(jtxtNombre1)
                             .addComponent(jtxtCapacidad, javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addComponent(jtxtID)
                             .addGroup(jPanel1Layout.createSequentialGroup()
                                 .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                                     .addComponent(jLabel6)
                                     .addComponent(jLabel7)
                                     .addComponent(jLabel8))
-                                .addGap(0, 0, Short.MAX_VALUE))
-                            .addComponent(jtxtID)))
-                    .addGroup(jPanel1Layout.createSequentialGroup()
+                                .addGap(0, 0, Short.MAX_VALUE))))
+                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel1Layout.createSequentialGroup()
                         .addContainerGap()
                         .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
                             .addComponent(jLabel1, javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jLabel2, javax.swing.GroupLayout.Alignment.LEADING)))
-                    .addGroup(jPanel1Layout.createSequentialGroup()
+                            .addComponent(jLabel2, javax.swing.GroupLayout.Alignment.LEADING))
+                        .addGap(0, 0, Short.MAX_VALUE))
+                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel1Layout.createSequentialGroup()
+                        .addGap(21, 21, 21)
+                        .addComponent(jLabel12, javax.swing.GroupLayout.PREFERRED_SIZE, 32, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(btnNuevo, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel1Layout.createSequentialGroup()
                         .addGap(19, 19, 19)
                         .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                             .addComponent(jLabel5)
@@ -201,15 +290,12 @@ public class Aulas extends javax.swing.JPanel {
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                         .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                             .addComponent(jbtnGuardar, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                            .addComponent(jbtnModificar, javax.swing.GroupLayout.DEFAULT_SIZE, 110, Short.MAX_VALUE)
+                            .addComponent(jbtnModificar, javax.swing.GroupLayout.DEFAULT_SIZE, 114, Short.MAX_VALUE)
                             .addComponent(jbtnEliminar, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                             .addComponent(jButton1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
-                    .addGroup(jPanel1Layout.createSequentialGroup()
+                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel1Layout.createSequentialGroup()
                         .addContainerGap()
-                        .addComponent(ComboTipo, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addContainerGap()
-                        .addComponent(ComboEdificio, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+                        .addComponent(comboTipo, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
                 .addContainerGap())
         );
         jPanel1Layout.setVerticalGroup(
@@ -230,12 +316,16 @@ public class Aulas extends javax.swing.JPanel {
                 .addGap(14, 14, 14)
                 .addComponent(jLabel7)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(ComboEdificio, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(20, 20, 20)
+                .addComponent(comboEdificio, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(18, 18, 18)
                 .addComponent(jLabel8)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(ComboTipo, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 105, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(comboTipo, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(46, 46, 46)
+                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(btnNuevo)
+                    .addComponent(jLabel12, javax.swing.GroupLayout.PREFERRED_SIZE, 34, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(18, 18, 18)
                 .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addComponent(jbtnGuardar)
                     .addComponent(jLabel3))
@@ -251,7 +341,7 @@ public class Aulas extends javax.swing.JPanel {
                 .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addComponent(jLabel9, javax.swing.GroupLayout.PREFERRED_SIZE, 24, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(jButton1))
-                .addContainerGap())
+                .addContainerGap(22, Short.MAX_VALUE))
         );
 
         jPanel2.setBackground(new java.awt.Color(255, 255, 255));
@@ -372,7 +462,7 @@ public class Aulas extends javax.swing.JPanel {
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 595, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, 611, Short.MAX_VALUE)
                     .addGroup(layout.createSequentialGroup()
                         .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addGap(30, 30, 30)
@@ -382,8 +472,9 @@ public class Aulas extends javax.swing.JPanel {
                                 .addComponent(busqueda, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
                             .addComponent(btnBuscar, javax.swing.GroupLayout.PREFERRED_SIZE, 19, javax.swing.GroupLayout.PREFERRED_SIZE))
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, 446, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                        .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, 446, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 22, Short.MAX_VALUE)))
+                .addContainerGap())
         );
 
         jPanel1.getAccessibleContext().setAccessibleName("Datos aulas");
@@ -400,19 +491,11 @@ public class Aulas extends javax.swing.JPanel {
       
              
         CRUDEAulas objetoAulas = new CRUDEAulas();
-        if (!"".equals(jtxtID.getText())) {
-           
-        }else{
-            if (objetoAulas.datosVacios(jtxtNombre1, jtxtCapacidad)==true) {
-                JOptionPane.showMessageDialog(null, "Por favor, ingresar todos los campos");
-            } else {
-                if (objetoAulas.espacioExistente(jtxtNombre1) == false) {
-                    objetoAulas.insertarAulas(jtxtNombre1, jtxtCapacidad, ComboEdificio, ComboTipo);
-                    objetoAulas.mostrarAulas(jtblListaAulas, busqueda.getText());
-                } else {
-                    JOptionPane.showMessageDialog(null, "El espacio ya se encuentra registrado");
-                }
-            }
+        if (objetoAulas.espacioExistente(jtxtNombre1)==false) {
+        objetoAulas.insertarAulas(jtxtNombre1, jtxtCapacidad, comboEdificio, comboTipo);
+        objetoAulas.mostrarAulas(jtblListaAulas,busqueda.getText());    
+        } else {
+            JOptionPane.showMessageDialog(null, "El espacio ya se encuentra registrado");
         }
          objetoAulas.mostrarAulas(jtblListaAulas, busqueda.getText());
         
@@ -423,21 +506,35 @@ public class Aulas extends javax.swing.JPanel {
     private void jbtnEliminarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jbtnEliminarActionPerformed
         // TODO add your handling code here:
         CRUDEAulas objetoAulas = new CRUDEAulas();
+
+    int respuesta = JOptionPane.showConfirmDialog(null, "¿Está seguro de que quiere eliminar este registro?", "Confirmación", JOptionPane.YES_NO_OPTION);
+
+    if (respuesta == JOptionPane.YES_OPTION) {
         objetoAulas.deleteAulas(jtxtID);
-        objetoAulas.mostrarAulas(jtblListaAulas,busqueda.getText());
+        objetoAulas.mostrarAulas(jtblListaAulas, busqueda.getText());
+    }
     }//GEN-LAST:event_jbtnEliminarActionPerformed
 
     private void jbtnModificarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jbtnModificarActionPerformed
         // TODO add your handling code here:
         CRUDEAulas objetoAulas = new CRUDEAulas();
-        objetoAulas.updateAulas(jtxtID, jtxtNombre1, jtxtCapacidad, ComboEdificio, ComboTipo);
-        objetoAulas.mostrarAulas(jtblListaAulas,busqueda.getText());
+
+    int respuesta = JOptionPane.showConfirmDialog(null, "¿Está seguro de que quiere modificar este registro?", "Confirmación", JOptionPane.YES_NO_OPTION);
+
+    if (respuesta == JOptionPane.YES_OPTION) {
+        objetoAulas.updateAulas(jtxtID, jtxtNombre1, jtxtCapacidad, comboEdificio, comboTipo);
+        objetoAulas.mostrarAulas(jtblListaAulas, busqueda.getText());
+        botonesEditar();
+        textosEditar();
+    }
     }//GEN-LAST:event_jbtnModificarActionPerformed
 
     private void jtblListaAulasMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jtblListaAulasMouseClicked
         // TODO add your handling code here:
         CRUDEAulas objetoAulas = new CRUDEAulas();
-         objetoAulas.SeleccionarEspacios(jtblListaAulas, jtxtID, jtxtNombre1, jtxtCapacidad, ComboEdificio, ComboTipo);
+         objetoAulas.SeleccionarEspacios(jtblListaAulas, jtxtID, jtxtNombre1, jtxtCapacidad, comboEdificio, comboTipo);
+         botonesEditar();
+         textosEditar();
     }//GEN-LAST:event_jtblListaAulasMouseClicked
 
     private void btnBuscarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnBuscarActionPerformed
@@ -453,6 +550,8 @@ public class Aulas extends javax.swing.JPanel {
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
         // TODO add your handling code here:
       limpiarCampos();
+      botonesInicio();
+      textosInicio();
 
     }//GEN-LAST:event_jButton1ActionPerformed
 
@@ -478,16 +577,26 @@ public class Aulas extends javax.swing.JPanel {
         
     }//GEN-LAST:event_busquedaKeyPressed
 
+    private void btnNuevoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnNuevoActionPerformed
+        // TODO add your handling code here:
+         botonesNuevo();
+        textosActivados();
+    }//GEN-LAST:event_btnNuevoActionPerformed
+
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JComboBox<String> ComboEdificio;
     private javax.swing.JComboBox<String> ComboTipo;
     private javax.swing.JButton btnBuscar;
+    private javax.swing.JButton btnNuevo;
     private javax.swing.JTextField busqueda;
+    private javax.swing.JComboBox<String> comboEdificio;
+    private javax.swing.JComboBox<String> comboTipo;
     private javax.swing.JButton jButton1;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel10;
     private javax.swing.JLabel jLabel11;
+    private javax.swing.JLabel jLabel12;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
     private javax.swing.JLabel jLabel4;
